Add `theme` support to `RadioGroup`
